requirements-test.txt
# Testing dependencies for iopool integration
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-aiohttp>=1.0.0
pytest-cov>=4.0.0
aioresponses>=0.7.0
homeassistant>=2024.1.0